pyproject.toml
[tool.poetry]
name = "corpus-base"
version = "0.0.3"
description = "Initialize justice, decision, citation, voting, and opinion tables."
authors = ["Marcelino G. Veloso III <mars@veloso.one>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.0"
citation-utils = "^0.0.9"
python-slugify = "^6.1.2"
sqlite-utils = "^3.29"
loguru = "^0.6.0"
markdownify = "^0.11.6"
unidecode = "^1.3.6"
httpx = "^0.23.0"
jinja2 = "^3.1.2"
python-frontmatter = "^1.0.0"

[tool.poetry.dev-dependencies]
rich = "^12.6"
pytest = "^7.1"
pytest-instafail = "^0.4.2"
pytest-sugar = "^0.9.4"
pytest-icdiff = "^0.5"
pytest-cov = "^2.12.1"
mypy = "^0.982"
black = "^22.10.0"
isort = "^5.10"
pre-commit = "^2.20"
ipykernel = "^6.12"
types-Markdown = "^3.4.0"
types-PyYAML = "^6.0.7"
types-python-slugify = "^5.0.4"
notebook = "^6.4.11"
jupyter = "^1.0.0"


[tool.pytest]
minversion = "7.1"
addopts = "-ra -q --cov=corpus_base tests/"
testpaths = ["tests"]

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 79
include_trailing_comma = 'True'

[tool.black]
line-length = 79
include = '.pyi?$'
exclude = '''
/(
    .git
    | .hg
    | .mypy_cache
    | .tox
    | .venv
    | _build
    | buck-out
    | build
    | dist
)/
'''


[build-system]
requires = ["poetry-core>=1.3.2"]
build-backend = "poetry.core.masonry.api"
